Create README.md
For a company to thrive, it is important to employ and retain talented employees. Some sources estimate that it costs up to a third of an employee’s salary when a worker leaves the organization, from recruiting, background checks, retraining, and interview time for a replacement as well as reduced productivity and lost knowledge from the departing employee (Agovino, 2019). It is therefore in an employer’s interest to reduce employee churn to increase their profitability.
To do this, they may wish to answer the following questions:
1)	How can employee turnover be reduced?
2)	Why do good employees leave and how can they be retained?
3)	Is it possible to predict or flag when and which employees are most likely to leave?